backend/requirements.txt
# 基础Web框架
Flask==3.0.0          # 最新稳定版，支持Python 3.8+
Flask-Cors==4.0.0     # 兼容Flask 3.0+的跨域支持
Requests==2.32.4    # 稳定版，支持HTTP请求
# 深度学习框架（CUDA 12.1适配，CPU用户可移除`+cu121`）
torch==2.3.1+cu121    # 最新PyTorch（根据CUDA版本调整，官网获取对应命令）
torchvision==0.17.0+cu121
torchaudio==2.6.0
transformers==4.46.2  # 最新版本，支持大模型和多种任务
# 机器学习库
scikit_learn==1.3.0
# 生物信息学核心
Bio==1.8.0  # BioPython核心模块（1.8.0版本，兼容性更好）
biopython==1.85       # 最新稳定版，增强序列处理功能
tqdm==5.81.0         # 进度条工具（较4.x版本更高效）
# 数据处理与可视化
numpy==1.26.0        # 稳定版（避免2.x版本兼容性问题）
pandas==2.1.1        # 最新功能增强版
matplotlib==3.8.2     # 修复高DPI显示问题，支持更多可视化特性
# API与外部服务
openai==1.72.0       # 最新官方SDK，支持ChatGPT API v1
python-dotenv==1.0.0  # 环境变量管理（推荐通过.env文件配置密钥）
# 知识库与向量检索
chromadb==0.5.23     # 向量数据库，保持与当前版本一致
sentence_transformers==3.2.1  # 文本向量化模型
PyPDF2==3.0.1        # PDF文件解析
python-docx==1.0.0   # Word文档解析
huggingface_hub==0.24.6  # Hugging Face 模型管理和下载
# 额外依赖（确保安装顺序）
pyparsing==3.1.0     # 解析库（matplotlib依赖）
huggingface_hub==0.26.2  # 拉取大参数模型